feat: support tail the logs from the victoriametrics database
optimize the code structure

we maybe need to tail the logs when we are debugging
